refactor: 패키지 명을 course 로 변경
