+++ src/main/java/me/chan99k/learningmanager/domain/course/Course.java
@@ -1,4 +1,4 @@
-package me.chan99k.learningmanager.domain.study;
+package me.chan99k.learningmanager.domain.course;
 
 import static org.springframework.util.Assert.*;
